+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file
+All notable changes to `unit-conversions` will be documented in this file
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2020-09-24
 
-- initial release
+- Initial release
+  - Convert kilograms to lbs feature.
+  - Convert celsius to kelvin feature.
+  - Convert celsius to Fahrenheit feature.